kagami/requirements.txt
PyYAML==3.11
six==1.6.1